@@ -11,7 +11,6 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
-        "axios": "^0.21.4",
         "bootstrap": "^5.1.3",
         "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
@@ -19,17 +18,18 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.14",
-        "vue-loader": "^15.9.7",
-        "vue-template-compiler": "^2.6.12"
+        "vue": "^3.2.33",
+        "vue-loader": "^17.0.0",
+        "vue-template-compiler": "^2.6.14"
     },
     "dependencies": {
         "autoprefixer": "10.4.5",
+        "axios": "^0.27.2",
         "moment": "^2.15.2",
         "vue-axios": "^3.4.1",
-        "vue-fragment": "^1.5.2",
         "vue-route": "^1.5.1",
-        "vue-router": "^3.5.3",
+        "vue-router": "^4.0.13",
+        "vuelidate": "^0.7.7",
         "watch": "^0.13.0"
     }
 }
