@@ -1,71 +1,31 @@
-// /**
-//  * First we will load all of this project's JavaScript dependencies which
-//  * includes Vue and other libraries. It is a great starting point when
-//  * building robust, powerful web applications using Vue and Laravel.
-//  */
-//
-// require('./bootstrap');
-//
-// window.Vue = require('vue').default;
-//
-// /**
-//  * The following block of code may be used to automatically register your
-//  * Vue components. It will recursively scan this directory for the Vue
-//  * components and automatically register them with their "basename".
-//  *
-//  * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
-//  */
-//
-// // const files = require.context('./', true, /\.vue$/i)
-// // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
-//
-// Vue.component('example-component', require('./components/ExampleComponent.vue').default);
-//
-// /**
-//  * Next, we will create a fresh Vue application instance and attach it to
-//  * the page. Then, you may begin adding components to this application
-//  * or customize the JavaScript scaffolding to fit your unique needs.
-//  */
-//
-// const app = new Vue({
-//     el: '#app',
-// });
-
-
 /**
  * First we will load all of this project's JavaScript dependencies which
  * includes Vue and other libraries. It is a great starting point when
  * building robust, powerful web applications using Vue and Laravel.
  */
-require('./bootstrap');
-window.Vue = require('vue').default;
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
-Vue.component('campaign-component', require('./components/CampaignComponent.vue').default);
-Vue.component('creative-component', require('./components/CreativeComponent.vue').default);
 
-// import App from './components/app.vue';
-import VueAxios from 'vue-axios';
-// import VueRouter from 'vue-router';
-import axios from 'axios';
-// import { routes } from './routes';
 /**
  * Next, we will create a fresh Vue application instance and attach it to
  * the page. Then, you may begin adding components to this application
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-// Vue.use(VueRouter);
-Vue.use(VueAxios, axios);
-//
-// const router = new VueRouter({
-//     mode: 'history',
-//     routes: routes
-// });
 
+
+require('./bootstrap');
+window.Vue = require('vue').default;
 
 import "../css/app.css"
-const app = new Vue({
-    el: '#app',
-    // router: router,
-    // render: h => h(App)
+import { createApp } from "vue"
+import VueAxios from 'vue-axios';
+import axios from 'axios';
+
+const app = createApp({
+
 });
+
+app.component('campaign-component', require('./components/CampaignComponent.vue').default);
+app.component('creative-component', require('./components/CreativeComponent.vue').default);
+
+app.use(VueAxios, axios);
+app.mount('#app');
